jwt validation

+++ src/controllers/UserController.ts
@@ -1,13 +1,14 @@
 import { PrismaClient } from "@prisma/client";
-import { IUser, PLogin, PUser, PUsername } from "types/ModelTypes";
+import bcrypt from "bcryptjs";
 import { hash } from "../utils/Hashing";
 import { FastifyRequest, FastifyReply } from "fastify";
 import { CreateUserInput, PasswordInput, UUIDInput, UserLoginInput, UsernameInput} from "schemas/UserSchema";
+import { Application } from "../App";
+
+import { isUserObject } from "../types/ModelTypes";
 
 const prisma = new PrismaClient();
  
-
-
 /*
 Type Casting Inputs -> Which to use, How to efficiently do it ?? 
 -- refer to schema page for refractoring then use the refractored.
@@ -19,7 +20,6 @@ export default class User {
     // args -> object of IUser but not saved in DB yet
     public static createUser = async (req: FastifyRequest<{Body: CreateUserInput}>, reply: FastifyReply) => {
         try {
-
             const {name, username, email, password} = req.body
             const hashedPassword = await hash(password);
             const user: CreateUserInput = await prisma.user.create({ data: {
@@ -37,30 +37,49 @@ export default class User {
 
     public static loginUser = async (req: FastifyRequest<{Body: UserLoginInput}>, reply: FastifyReply) => {
         try {
-            // unhash and save in session ? 
+            const {email, password} = req.body;
+            console.log(email, password);
+            const user = await prisma.user.findUnique({
+                where: {
+                    email: email
+                }
+            });
+            console.log(user);
+            if (user && await bcrypt.compare(password, user.password)) {
+                const accessToken = Application.jwt.sign(user, {expiresIn: "5m"});
+                console.log(accessToken)
+                reply.send({accessToken: accessToken})
+            } else {
+                console.log(`${email} does not exist/password wrong`);
+            }
         } catch (error) {
-            console.log(error);
+            console.log(error); 
         }
     }
 
-    public static updatePassword = async (req: FastifyRequest<{Params: UUIDInput , Body: PasswordInput}>, reply: FastifyReply) => {
-        try {
-            const {password} = req.body;
-            const {id} = req.params;
-            const updatedHashedPassword = await hash(password);
-            const updatedUserPassword = await prisma.user.update({
-                where : {
-                    id: id
-                }, 
-                data: {
-                    password: updatedHashedPassword
+    public static updatePassword = async (req: FastifyRequest<{Body: PasswordInput}>, reply: FastifyReply) => {
+        try {   
+                if (isUserObject(req.user)) {
+                    const {password} = req.body;
+                    const id = req.user.id
+                    const updatedHashedPassword = await hash(password);
+                    const updatedUserPassword = await prisma.user.update({
+                        where : {
+                            id: id
+                        }, 
+                        data: {
+                            password: updatedHashedPassword
+                        }
+                    })
+                    if (updatedUserPassword) {
+                        console.log(`---successfully updated password for ${id}---`)
+                        reply.status(201).send(updatedUserPassword)
+                    } else {
+                        console.log("update password unsuccessful")
+                    }
                 }
-            })
-            if (updatedUserPassword) {
-                console.log(`successfully updated password for ${id}`)
-            } else {
-                console.log("update password unsuccessful")
-            }
+
+                
         } catch (error) {
             console.log(error)
         }
@@ -68,24 +87,25 @@ export default class User {
     }
 
     
-    public static updateUsername = async (req: FastifyRequest<{Params: UUIDInput, Body: UsernameInput}>, reply: FastifyReply) => {
+    public static updateUsername = async (req: FastifyRequest<{ Body: UsernameInput}>, reply: FastifyReply) => {
         try {
-            console.log(req.params, req.body);
-            const {id} = req.params;
-            const {username} = req.body;
-            const updatedUsername = await prisma.user.update({
-                where: {
-                    id: id
-                },
-                data: {
-                    username: username
+            if (isUserObject(req.user)) {
+                const id = req.user.id
+                const {username} = req.body;
+                const updatedUsername = await prisma.user.update({
+                    where: {
+                        id: id
+                    },
+                    data: {
+                        username: username
+                    }
+                })
+                if (updatedUsername) {
+                    console.log(`successfully updated username to ${username}`)
+                    reply.send(updatedUsername)
+                } else {
+                    console.log("update username unsuccessful");
                 }
-            })
-            if (updatedUsername) {
-                console.log(`successfully updated username to ${username}`)
-                reply.send(updatedUsername)
-            } else {
-                console.log("update username unsuccessful");
             }
         } catch (error) {
             console.log(error)
@@ -93,39 +113,52 @@ export default class User {
     }
 
     // deleting in order of foreign key constraints 
-    public static deleteUser = async (req: FastifyRequest<{Params: UUIDInput}>, reply: FastifyReply) => {
+    public static deleteUser = async (req: FastifyRequest<{}>, reply: FastifyReply) => {
   
         try {
-            const {id} = req.params
-            const deleteReviews = prisma.reviews.deleteMany({
-                where: {
-                    userId: id
-                }
-            });
-            const deleteFollowing = prisma.follows.deleteMany({
-                where: {
-                    userId: id
-                }
-            })
-            const deleteFollowers = prisma.follows.deleteMany({
-                where: {
-                    friendId: id
-                }
-            });
-            const deleteUser = prisma.user.delete({
-                where: {
-                    id: id
-                }
-            })
-            await prisma.$transaction([deleteReviews, deleteFollowing, deleteFollowers, deleteUser])
-            console.log(`successfully deleted all tracers of ${id}`)
+            if (isUserObject(req.user)) {
+                const id = req.user.id;
+                const deleteReviews = prisma.reviews.deleteMany({
+                    where: {
+                        userId: id
+                    }
+                });
+                const deleteFollowing = prisma.follows.deleteMany({
+                    where: {
+                        userId: id
+                    }
+                })
+                const deleteFollowers = prisma.follows.deleteMany({
+                    where: {
+                        friendId: id
+                    }
+                });
+                const deleteUser = prisma.user.delete({
+                    where: {
+                        id: id
+                    }
+                })
+                await prisma.$transaction([deleteReviews, deleteFollowing, deleteFollowers, deleteUser])
+                console.log(`successfully deleted all tracers of ${id}`)
+            }
+
         } catch (error) {
             console.log(error)
         }
   
     }
 
-    
+    public static getUser = async (req: FastifyRequest<{}>, reply: FastifyReply) => {
+        try {
+            if (isUserObject(req.user)) {
+                reply.send(req.user)
+            } else {
+                console.log("idk why the code will even make it here? because validation will settle")
+            }
+        } catch (error) {
+            console.log(error)
+        }
+    }
 
     public static follow = async () => {
